backend-service/rag-service/requirements.txt
fastapi
uvicorn
supabase
sentence-transformers
transformers
torch
numpy
python-dotenv
psutil